requirements.txt
alembic==1.11.1
anyio==3.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.96.0
filelock==3.12.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2==2.9.6
pydantic==1.10.8
pytest==7.3.1
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.6.3
urllib3==2.0.2
uvicorn==0.22.0
